requirements.txt
Django
djangorestframework
psycopg2-binary
django-environ
Faker
drf-yasg
django-filter